requirements.txt
faiss==1.7.1
imageio==2.9.0
matplotlib==3.4.3
numpy==1.21.2
opencv_python==4.5.3.56
Pillow==9.4.0
sacred==0.8.2
scikit_image==0.18.3
scikit_learn==1.2.1
skimage==0.0
tensorboardX==2.5.1
torch==1.7.1
torchvision==0.8.0a0
tqdm==4.62.3
